BUG: Fix degenerate nans
If the mean of a particular landmark point was the origin for some
iteration, the result from then on would become all nans due to a divide
by zero.

Updating to just not divide by zero and skip using the landmark with
the mean of the origin for that iteration.
The mean will likely not be the origin the next iteration because the
rotation will be different.